Refactored YAML editing to better support parameters
We needed a way to support arbitrarily complex parameters in both
type and nesting level. As a temporary solution I enabled the user
to toggle between a key value component and a yaml editor.

To enable this I refactored ResourceYaml into a composable YamlEditor
component that can be used elsewhere.

rancher/dashboard#388
rancher/dashboard#380