Disable editing a constraint's constraint template
Editing a constraint's constraint template would require us to delete
the existing constraint and create a new one transactionally. Since
the backend doesn't support this transaction of we're disabling the ability
to edit which template is being used.

I also noticed a number of inconcictencies and poor behaviors
with the LabeledSelect component so I went ahead and made
it consistent with LabeledInput.
- The search field switched to a color distinct from the rest of the container
  when disabled and hovered.
- There wasn't a focus outline.
- The size was slightly different because it was in a container
- There was a distinct background color around the arrow indicator

rancher/dashboard#433