App model specialization and OPA Gatekeeper menu items
When working on adding 'Add Constraint' and 'Add Template' to the
OPA Gatekeeper page I noticed that there wasn't a good way to
create custom models for specific chart app types. To fix this I went
ahead and made it so that we can load app models from the
'models/apps' directory and merge that with the 'models/app' model.
I think this will be useful for future apps like istio.

rancher/dashboard#391