Correctly wait for the gatekeeper app to be ready
app doesn't have a state so all the logic was a little wonky, the state must
have been removed. setup the app to watch for the deployed condition instead.
refactored the logic about sending out the emit a bit too so that the config
observer sends the event only if the wait for contidion call int he enable has
failed. it seems a little bonkers to wait 60 seconds which can happen some
times. 30 second default should be more than enough in most cases so this really
catches outside cases where clusters may be bogged down and slow.

rancher/dashboard#371 - Scenario 2