Change DetailTop to use css grid
flex box is nice for growing and shrinking content but in somecases the wrapping
is really off. Css grid works better in detail top because we can align the
content more simply and we can define a set number of columns. Additionally with
media queries we can easily change the amount of columns on the fly.