Remove custom VueRouter.resolve
replace and push both work  as intended. You can
specify a a full url  if you don't want to have the baseUrl
prepended. Better yet, if you're navigating within
the app it's better to specify a location.

The custom resolve was actually breakinging a huge
amount of links in the app which was caused 404s if a
user attempted to open a link in a new tab. Most things appeared to work presumably because vue-router
was handling the page switches instead of the browser.

rancher/dashboard#643