Fix clone as yaml
The YAML that was generated for the 'Clone as Yaml' page had several
fields that shouldn't be present when cloning. A cleanForNew method
was already being invoked on the model so it seemed better to convert
the clean model to yaml instead of requesting the backend yaml from the
original model.

rancher/dashboard#641