Adding pathType to ingress rules
Creating an ingress in k8s 1.19 was failing because we didn't
have pathType. This adds the new option as an InputWithSelect.
Because this field is present in 1.18 I'm showing this field for
all implementations.

I also fixed a small issue under DefaultBackend where
selecing a TargetService was removing the port label. The
label field was just missing from the LabeledSelect component.

rancher/dashboard#1634
https://github.com/rancher/dashboard/issues/1301#issuecomment-705052366