refactored how cluster switcher is positioned on screen
because cluster switcher is on the right edge of the screen and the content of the dropdown is auto width, the dropdown would appear off screen with long cluster names.
Added the filter method to check for zero options and to update the position of the dropdown with the popper update method.