Make the capacity of a PVC match a PV when selected
When creating a PVC from an existing PV ember would set the capacity of the PVC to be the capacity of the selected PV. This adds that same behavior.

As a part of this I reset the value of the pv when switching sources to be consistent and to force the user to select the pv in order to trigger setting the capacity in case the value was changed by switching between sources.

rancher/dashboard#2448