Add containerStatus lastState information to Pod detail page
The containerStatus object (https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.22/#containerstatus-v1-core) also contains a lastState field which gives you details about the container's last termination condition.

If there's a lastState in the containerStatus, this should also be displayed in the container list.

This is especially useful when you need to debug why a container restarted, because it gives you the exit code, potential error message and reason as well as the time when the container terminated.

Addresses https://github.com/rancher/dashboard/issues/6189

Signed-off-by: Bastian Hofmann <bashofmann@gmail.com>