Create e2e test for `configmap`
This test allows us to assert that blocks containing long lines (over 80 characters) entered into a configmap value field will output the correct value upon saving. The intent of this test is to assert that the regression described in #10044 is resolved and will not be reintroduced if further changes are made to `create-yaml`.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>