Replace `Vue.extend` with `defineComponent` in `LabeledInput.vue`
Rewrote `compact-input.ts` and `labeled-form-element.ts` mixins as composables to address TypeScript issues when using mixins with `defineComponent`. The new composables are written to adhere to the conventions and best practices [^1] as best as possible, taking into consideration what's available to us in Vue 2.7.14.

Moving to composables positions us for a smoother Vue 3 migration.

[^1]: https://vuejs.org/guide/reusability/composables.html#conventions-and-best-practices

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>