Modify bottom margin for `AddCustomBadgeDialog` title
This adjusts the margin for the title of `AddCustomBadgeDialog.vue` to give the appearance of centering. The design of the card component makes it difficult to give a nice centered look with the use of padding around the entire card component, in addition to the horizontal rule. This is a bit of a hack to help without modifying portions of card.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>