Making sure tabTitle updates if the content of a slot updates
Watches and computed props didn't update based on the content of slots.

The render method would still be called when the slot was updated so I moved the update to there. I also just use a method to compute the title because the value would get cached and not updated even if you accessed it more than once.

https://github.com/rancher/dashboard/issues/9822#issuecomment-1976814164