Initialize i18n after plugins load more
We previously initialized i18n on every middleware invocation. This happened to be called after harvester was loaded by the authenticated middleware which loaded it's i18n keys.

This will now call i18n/init after the plugin locales have been loaded. This will be non-blocking so I don't see it as being an issue that we call it once for each plugin.