Changing the way we handle redirects by two helper functions inside authenticated middleware
We didn't actually need to call redirect from the methods since the `loadingError` store action did the redirect for us.

We still need to exit early from the middleware so instead of having to check and execute a return value (that was a boolean so couldn't be executed anyway) we now just throw an exception.