Fixing harvester cluster page reload 404s
We were previously calling authenticated middleware to handle the loading of the harvester plugin. This will now explicitly do the loading instead of relying on the middleware.

This showed up again because of two factors. We were no longer executing the authenticated middlware because of our guards at the top and because the route.currentRoute.path was showing '/' instead of the full path.