Updating how we render custom list pages
Vue3 changes how attributes are inherited https://vuejs.org/guide/components/attrs#attribute-inheritance

This addresses the issue by adding a `inheritAttrs: false` to the resource-fetch mixin which seems to correlate with the usage of ResourceTable as a root element. Specifically what happened is that when RootTable was used at the root in these cases the value of :rows would get overwritten causing the table to be empty.

I also fixed some spacing issues by changing how we render "outlet" also an inherritedAttr.