Refactoring tabbed values in chart installation
We had this `hasTabs` api which implicitly allowed a chart install override component to implicitly define tabs instead of just having the component define them explicitly.

This required prop drilling to work and it doesn't appear to work any longer with dynamic components. Since this was only used in two places I removed support for this and simplified the implementation.