Adding in some fixes for extensions in vue3
The first one we needed a way to include the vue library for the extensions which are built as libraries and therefore do not include vue as a part of the bundle.

The second one was related to routes. The old way that we replaced the router matcher no longer seems to function. I'm guessing that the router methods are closures which are capturing a matcher that we don't actually see. We may want to come up with a faster way to update routes.