Clone `defaultContainer` when applying defaults
`defaultContainer` retains its values even after navigating away from the workload create form. This behavior was not observed in Vue2, and it requires further investigation to fully understand.

To address this issue in the short term, using `structuredClone` to create a deep copy of defaultContainer ensures that modifications to the container do not persist after navigating away from the form.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>