Fix cloud-credential expiry display
Fix cloud-credential being displayed if the expiry annotation has a Unix
timestamp of "0".

When Rancher is upgraded and there are cloud-credentials for a Harvester
cluster without a token (e.g. when the token expired and purged or
was manually removed), the cloud-credential's expiry will be updated
with a Unix timestamp of "0" (i.e. as if the token had expired at 00:00
AM on Jan 1st 1970), since the actual time of expiry of the token can
not be determined.
This fix makes the UI display such a cloud credential correctly as
"expired".

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>