Fix cloud credential sort order by expiry date
Fix sort oder of cloud credentials if sorted by expiry date.

If the annotation `rancher.io/expiration-timestamp` is `"0"`, indicating
that the token for that cloud credential has long expired and/or was
deleted, when the migration from Rancher <= v2.9.2 to >= v2.9.3
happened, then the cloud credential should show up as expired. It should
also be sorted among the other expired credentials, if the list is
sorted by expiration state.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>