Avoid embedding web-worker polyfill
The BLOB code from web-worker is minified, and contains source maps that are then
passed over by webpack - that creates a situation of minified code inside BLOB
inside minified code which confuses some tools, eg.

https://github.com/jantimon/chrome-profile-sourcemap-resolver/pull/4

That could be fixed, but the only reason to do so would be supporting very old
(pre-2013) browsers, so we are better off removing the polyfill altogether.

Signed-off-by: Silvio Moioli <silvio@moioli.net>