Use `semver.coerce()` to get valid kubernetes version
When creating AKS clusters, `1.29` is a valid kubernetes version that can be supplied to provision a cluster with the latest patch version. Unfortunately, `aksConfig?.kubernetesVersion` only stores `1.29` and the fully compatible semantic version is unknown. With the information we have available to us, we can utilize `semver.coerce()` to at least attempt to infer a Kubernetes version with the information that we have available to us.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>