Strip origin when obtaining `url` in `urlFor()`
There are instances where Steve can provide the url of the API server for content links (e.g. `http://127.0.0.1:9443/v1/management.cattle.io.clusters`); this can prove to be problematic if the URL for Steve and the frontend server differ, resulting in CORS errors when Dashboard attempts to communicate with Steve directly. By stripping the origin from the URL, we can ensure that uses the correct base url and always routes API traffic through a proxy.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>