Remove `onChanges()` method in `ResourceYaml.vue`
This removes the `onChanges()` method from `ResourceYaml.vue` because the intent of the code is not clear and appears to provide no benefit in testing. This is a source of frustration for some users, making it impossible to add comments to YAML. 

It appears that this function exists to automatically adjust the indentation of comment lines; it has failed to accomplish this task during testing, so it appears to be meaningless. If users wish to format yaml, we should support this with an explicit action in the YAML editor, not something that is performed on every keypress.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>