Disabled failing AKS test
The test failure associated with the test "should display subnets grouped by network in the virtual network dropdown" is legitimate and reproducible in master - it is just that the unit test was not detecting this error. I suspect that bumping Vue and related dependencies has patched a bug in test utils.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>