CloudCredential Model: Fix CC expiry display
Fix CloudCredential expiry date not being displayed if it's more than a
week in the future.
In this case, the behavior was the same as if there wasn't an expiry
date at all. This was confusing to users.
Fix it by always displaying the expiry date for CloudCredentials if it's
set.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>