SortableTable: Emit an `enter` event for table selection
This refactors the event listener that was attached to the window so that it is now attached the sortable table. The table needs to listen to the keyup event and the row needs to receive focus in order for this to work; the row's tabindex is set to -1 so that it is not navigable via keyboard.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>