Sanitize the version when setting it in UIPlugin label (#13585)
This change enhances the extension-server.labels directive in _helpers.tpl
to avoid error when setting the app.kubernetes.io/version label with version
which includes + character. This is valid for semver but not for label version.

The solution adopts the convention used in Helm which replaces + with _ when
storing the version in places which don't allow + usage, e.g. OCI repository.

The same approach is already used in the `extension-server.chart` directive,
which is also used as a label value.

Fixes Issue 13584