Revert Vue 3.5 features added to dashboard to maintain compatibility with Vue 3.2 (#13880)
* Replace `useTemplateRef` with `ref` for Vue 3.2.47 compatibility

`useTemplateRef` was added to Vue 3.5, but cannot be used in extensions that might be installed in Rancher 2.10 because that has a dependency on Vue 3.2.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Replace Reactive Props Destructure with `withDefaults` macro

Reactive Props Destructure was added to Vue 3.5, but cannot be used in extensions that might be installed in Rancher 2.10 because that has a dependency on Vue 3.2.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Add feature flag to conditionally render dropdown menu in tables

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

---------

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>