Add ability to handle objects to `mergeWithReplaceArrays`
This renames `mergeWithReplaceArrays` to `mergeWithReplace` and adds functionality for working with objects. I'm opting to extend the existing util after I found that objects are not the only concern In reviewing prior work #14165 that dealt with arrays. Defaulting `replaceObjectProps` to `false` keeps the existing behavior in  place for other usages of `mergeWithReplace`, limiting the chance for regressions in rke2 yaml.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>