Overwrite default chart values with user supplied values (#14200)
* Overwrite default chart values with user supplied values

Changes to merge behavior so that user supplied values overwrite default values when merging. The lodash merge operation merges object properties recursively, meaning that any keys in the left-most object will remain present. The customizer function works to overwrite any duplicate properties present in the user values.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Add ability to handle objects to `mergeWithReplaceArrays`

This renames `mergeWithReplaceArrays` to `mergeWithReplace` and adds functionality for working with objects. I'm opting to extend the existing util after I found that objects are not the only concern In reviewing prior work #14165 that dealt with arrays. Defaulting `replaceObjectProps` to `false` keeps the existing behavior in  place for other usages of `mergeWithReplace`, limiting the chance for regressions in rke2 yaml.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Unit test object merging behavior

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

---------

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>