Prevent default form submit behavior for open ldap
This prevents the default form submit behavior in `ldap.vue`.

 #13954 introduced a change to resolve issues with focus behavior when the async button component entered a disabled state. It turns out that the disabled attributed on the button was preventing the default submit behavior, and changing this reintroduced the default form submit behavior.

The recommended way to submit forms in vue is to utilize `@submit.prevent` to disable the default form submit behavior and define our own submit logic. We want to retain the changes made to async button so that keyboard focus remains consistent.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>