Revert custom object merge behavior
Revert changes to object merge behavior because the existing behavior is destructive when merging deeply nested user values. For example, a default object contains the value:

```js
a: {
  b: true,
  c: true,
  d: {
    e: true,
    f: {
      g: true,
      e: { },
      f: [ ],
    },
  },
}
```

and a user value modifies `g`, which is stored as:

```js
a: {
  d: {
    f: {
      g: false,
    },
  },
}
```

When the object is merged, deeply nested values will only be taken from the user values, causing properties to be lost. Furthermore, default properties for `f` can also be dropped. This operation is too destructive, and it's better to merge user-supplied properties with defaults than to drastically alter YAML altogether.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>