Set CodeMirror `readonly` flag to `true` to enable copy
This updates the `readonly` flag to `true` instead of `nocursor` when the CodeMirror component is in View mode. The copy event will not fire when type `ctrl + C` or `Command + C`; this stems from the fact that the editor never receives focus when `nocursor` is used.

To resolve this issue, we can use a boolean value for `readonly` and ensure that the container focus is not set to true when in View mode to prevent help text from rendering.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>