Update value watcher for ArrayList
The previous implementation attempted to preserve the id when newVal matched oldVal. This means that a new id would be generated when editing an existing value, causing the component to re-render and lose focus on the input.

This new approach preserves the existing id by locating the modified index and updating the value in place.

This solution comes with a tradeoff - add or deleting values will cause a new ID to be generated for each member of the ArrayList. This will force a re-render of the entire list and can become expensive when the ArrayList becomes arbitrarily large. Attempting to preserve the ids can become complicated on deletion when multiple members of the ArrayList have the same value.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>