[Backport 2.11.4] Enter edit mode if updating roles fails during user creation (#14765)
* Enter edit mode if updating roles fails during user creation

This emits  an `update:mode` event that sets to mode to "edit" and refreshes all resource-related variables when the `updateRoles()` method fails during user creation. This provides an opportunity for users to correct mistakes while still staying in the "Create User" context.

An alternate approach that explored navigating to an edit route was explored, but this required re-rendering the entire form and would eliminate the error message; additional state to persist the error messages would be required and I found that requirement to be more complex than the approach taken here.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Display a user readable error an escalation error is raised

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* e2e test: attempt to create an admin user with standard user

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Prevent watcher from overriding values when user id changes

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

* Generate new timestamp for admin user

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>

---------

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>