Update charts e2e test so that it doesn't rely on version
This updates the `should call fetch when route query changes with valid parameters` test so that it navigates to the first active link in the list of versions.

The test does not need to rely on a specific version in order to assert its case. Relying on the specific version is prone to failure if the backend stops providing it for any reason.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>