Update secret path for port-pr workflow
This reverts the change to the secrets for the `port-pr` workflow.

The original attempt to read secrets in 601bbb30bb465e5d1a161a40c9b4ede75b22931c omitted the permissions entirely. This could have been the contributor as to why we could not get secrets from the vault on the first attempt.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>