Allow for custom class attrs in selector buttons
This change allows a class of `role-primary` for the default in `shell/components/form/FileImageSelector.vue`. The `btn` class is dropped because it is redundant; `btn` is already defined in the child `FileSelector` component.

The previous implementation would merge classes in such a way that both `role-primary` and `role-secondary` classes would be present on a button.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>